==> _episodes_rmd/06-equate.Rmd <==
---
title: "Part 2: Test Equating"
teaching: 30
exercises: 20
questions:
  - "How do I prepare data for test equating?"
  - "How do I conduct test equating?"
  - "How do I extract indices of interest for comparison, reporting, and analysis?"
  - "how do I visualize equated relationships?"
objectives:
   - "Prepare frequency tables of test scores."
   - "Use `equate` to conduct equating studies."
   - "Extract estimates of error for presenting in figures."
keypoints:
   - "`equate` can be used for small- and large-sample test equating."
---

```{r, include=FALSE}
source("../bin/chunk-options.R")
knitr_fig_path("06-")
source("../bin/download_data.R")
```

We start by loading the required packages.

```{r load-package, message=FALSE, purl=FALSE}
library(tidyverse)
library(equate)
```

If not still in the workspace, load the data we saved in the previous lesson. We also need to
calculate the raw total and raw anchor scores for the reading and listening tests,


```{r load-data, purl=FALSE, warning=FALSE, message=FALSE, error=FALSE}
# load test_1_raw_totals
test_1_raw <- read_csv('data_output/test_1_raw_totals.csv') %>% # read in our total score data from earlier
  select(-country)

# load placement 1 and 2 data
test_results_1 <- read_csv('data/placement_1.csv')

test_results_2 <- read_csv('data/placement_2.csv')

# create total raw and anchor scores
test_2_raw <- test_results_2 %>%
  mutate(raw_total = rowSums(.[5:69], na.rm = TRUE)) %>%
  mutate(read_raw_total = rowSums(select(., contains("_read_")), na.rm = TRUE)) %>%
  mutate(list_raw_total = rowSums(select(., contains("_list_")), na.rm = TRUE)) %>%
  mutate(list_an_raw = rowSums(select(., matches("q\\d+_list_\\w{2,4}_an")), na.rm = TRUE)) %>% # ugly regex for summing listening anchor items
  mutate(read_an_raw = rowSums(select(., matches("q\\d+_read_\\w{2,4}_an")), na.rm = TRUE)) %>%
  select(ID, contains('raw'))

# write_csv(test_2_raw, 'data_output/test_2_raw_totals.csv')
```

## Preparing the data

The `equate` function requires frequency tables of the test scores that are to be equated. We
are working with data that was collected under the Non-equivalent group anchor test (NEAT) design.
As a result, we need bivariate frequency tables, or frequency tables of score combinations on the 
anchor and total test. The `equate` package has a function named `freqtab` to help us compute these.
We will be using two arguments in this function: `x` and `scales`. We will start with the listening 
test.

One quick way for us to figure our how many items are on the total test and how many are on the 
anchor test is to `select` the columns of the items and then count them:

```{r, purl=FALSE}
listen_1_total_q <- test_results_1 %>%
  select(contains('list'))

listen_1_an_q <- test_results_1 %>%
  select(contains('list')) %>%
  select(contains('an'))

ncol(listen_1_total_q)
ncol(listen_1_an_q)
```

```{r, purl=FALSE}
listen_1_freq <- freqtab(test_1_raw[c('list_raw_total', 'list_an_raw')], scales = list(0:35, 0:9))
```

> ## Exercise
>
> Figure out the total and anchor scales for the `test_2_raw` data and then create
> a frequency table for it.
>
> > ## Solution
> >
> > ```{r equate-freq, answer=TRUE, purl=FALSE}
> > # How many questions are on the total and anchor forms?
> > listen_2_total_q <- test_results_2 %>%
> >   select(contains('list'))
> > 
> > listen_2_an_q <- test_results_2 %>%
> >   select(contains('list')) %>%
> >   select(contains('an'))
> > 
> > ncol(listen_2_total_q)
> > ncol(listen_2_an_q)
> > 
> > # Create the frequency table
> >
> > listen_2_freq <- freqtab(test_2_raw[c('list_raw_total', 'list_an_raw')], scales = list(0:30, 0:9))
> > ```
> {: .solution}
{: .challenge}


Now that we have the frequency tables, we are ready to equate. There are a number of consderations
in choosing an equating method [Kolen & Brennan's _Test Equating, Scaling, and Linking_](https://www.amazon.com/Test-Equating-Scaling-Linking-Statistics/dp/1493903160)
is a nice resource for those of you who want to take a deep dive into it.

We will use the circle-arc method right now to equate form 1 on to scores from form 2:

```{r, purl=FALSE}
list_ca <- equate(listen_1_freq, listen_2_freq, type = 'circle-arc', lowp = c(0, 0), highp = c(35, 30))
```


> ## Exercise
>
> What does a score of 32 on form 1 concord to on form 2?
> (hint: `str` or `glimpse`)
>
> > ## Solution
> >
> > ```{r equate-conc, answer=TRUE, purl=FALSE}
> > list_ca$concordance
> > ```
> {: .solution}
{: .challenge}

There are three types of error associated with equating: random error, systematic error (or bias),
and total error. To calculate the latter two, we would need a criterion, or gold standard, equating
relationship between the two test forms that we could compare our circle-arc methods to. We dont have
that, so we will estimate the standard error by bootstrapping the equating relationship. One way to 
inspect the random error is by plotting it:

```{r, purl=FALSE}
list_ca_see <- bootstrap(list_ca, reps = 100)

list_ca_see$se

plot(list_ca_see, out = 'se')
```

Usually, in carrying out a full equating study, multiple equating relationships are estimated and compared. Below is a demonstration of how this can be done in a few lines of code.

```{r}

neat_args <- list(identity = list(type = "identity"),
                  mean_tuck = list(type = "mean", method = "tucker"),
                  mean_nomi = list(type = "mean", method = "nominal weights"),
                  line_tuck = list(type = "linear", method = "tucker"),
                  line_chai = list(type = "linear", method = "chained"),
                  circ_tuck = list(type = "circle-arc", method = "tucker"),
                  circ_chai = list(type = "circle-arc", method = "chained", chainmidp = "linear"))

comp_meth <- bootstrap(x = listen_1_freq, y = listen_2_freq, reps = 100, args = neat_args)

plot(comp_meth, out = "se", addident = FALSE, legendplace = 'top')

round(summary(comp_meth), 2)
```

```{r, fig.width=12}
# using ggplot and dplyr/tidyr

fiver <- comp_meth$se %>%
  as_tibble() %>%
  select(-identity) %>%
  mutate(score = 0:35) %>%
  slice(seq(1, 36, by = 5)) %>%
  gather(key = method, value, -score)

comp_plot <- comp_meth$se %>%
  as_tibble() %>%
  select(-identity) %>%
  mutate(score = 0:35) %>%
  gather(key = method, value, -score) %>%
  ggplot(., aes(x = score, y = value, colour = method, shape = method)) +
  geom_point(data = fiver, aes(size = value), show.legend = FALSE) +
  geom_line() +
  theme_gray(base_size = 18) +
  scale_colour_viridis_d() +
  scale_x_continuous(breaks = c(0, 5, 10, 15, 20, 25, 30, 35), limits = c(0, 35)) +
  theme(panel.grid.minor = element_blank())

comp_plot

comp_table <- comp_meth$se %>%
  as_tibble() %>%
  summarise_all(., 'mean') %>%
  mutate_all(., 'round', 2)

comp_table

```



{% include links.md %}
