==> _episodes_rmd/11-office-two.Rmd <==
---
title: "Part 6: Office Hours Part Two"
teaching: 60
exercises: 35
questions: 
  - "Can I apply anything I've learned today to new data?"
objectives: 
  - "Read in the data"
  - "Select the Listening Test items"
  - "Describe the distribution of total scores"
  - "Provide an estimate of CTT reliability"
  - "Conduct a Rasch analysis of the Listening Test data"
  - "Summarize the Person and Item measurement facets (i.e., average ability/difficulty, SE)"
  - "Calculate Person Separation"
  - "Produce a Wright Person-Item map"
  - "Produce Rasch item statistics in a table"
  - "Include all of the above in a technical report in PDF format"
keypoints: 
  - "You can apply stuff that you learned today!"
source: Rmd
---

```{r, include=FALSE}
source("../bin/chunk-options.R")
knitr_fig_path("12-")
source("../bin/download_data.R")
```

Day 2 Problem Set

HINT: You can start working in a standard R script. Then, once your analyses and plots
look fine, start a new R Notebook and begin laying things out - you can copy your code
from the script and place it in chunks in your R Notebook. Or you can read it into your 
.Rmd document at the top using `source` and then call on the objects you want to print
in your chunks.

You can download a prepped document or start your own.

```{r, purl=FALSE, eval=FALSE}
download.file("https://raw.githubusercontent.com/gtlaflair/ltrc-2019/gh-pages/documents/listening-test-analysis.Rmd",
              "documents/listening-test-analysis.Rmd", mode = "wb")
```

You can download our answers here.

```{r, purl=FALSE, eval=FALSE}
download.file("https://raw.githubusercontent.com/gtlaflair/ltrc-2019/gh-pages/scripts/day-2-problem-set-solution.R",
              "scripts/day-2-problem-set-solution.R", mode = "wb")
```

{% include links.md %}
